backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
jinja2
python-multipart
